DHbio.html
Bio for Donovan. When people send me their Bios, or if you just want to make your own Bio page, You can just copy the code for this one, and fill in the information (that's what I was going to do anyway). If there are finer details that you want on your specific page, we can make those changes together maybe? Anyways, this is this basic structure of the bio page. The lesson pages I imagined being fairly empty with the video in it, and a brief description (with the CSS document and/or whatever Nelly researches to make the website look better).